refactor: Remove unnecessary codes
