+++ lib/customTypes/ScheduleTypes.ts
